Move units into a top level folder
