+++ emergence_lib/src/units/item_interaction.rs
